+++ ivy/functional/backends/numpy/data_type.py
@@ -226,3 +226,10 @@ def dtype_bits(dtype_in: Union[np.dtype, str], /) -> int:
         .replace("float", "")
         .replace("complex", "")
     )
+
+
+def is_native_dtype(dtype_in: Union[np.dtype, str], /) -> bool:
+    if dtype_in in ivy_dtype_dict:
+        return True
+    else:
+        return False
