--- ivy/functional/backends/paddle/experimental/data_type.py
@@ -1,15 +1,0 @@
-# global
-from typing import Union
-
-import paddle
-from ivy.utils.exceptions import IvyNotImplementedException
-
-# local
-from ivy.functional.backends.paddle import ivy_dtype_dict
-
-
-def is_native_dtype(dtype_in: Union[paddle.dtype, str], /) -> bool:
-    if dtype_in in ivy_dtype_dict:
-        return True
-    else:
-        return False
